Extract span metrics to a dependency-free module
span_eval_metrics.py (pure stdlib) holds char_set/span_metrics/
print_metrics_table so both the encoder eval and the vLLM generative
eval import the identical metric without pulling torch/lettucedetect
into the vLLM env.
